finished authentication. testing

+++ secondary-modify.php
@@ -59,7 +59,7 @@
 
 					<div class="rdmore">
 						<a class="read-more" href="https://<?php echo $data["link"];?>">Read More!</a>
-
+ <?php if(isset($_SESSION['username'])) { ?>
 						<div>
 							<a class = "edit" href="edit.php?id=<?php echo $data['id']; ?>">Edit</a>
 							<a class = "delete" onclick="return confirm('Are you sure you want to delete: <?php echo $data["firstname"] . " " . $data["lastname"]; ?>?')" href="delete.php?id=<?php echo $data['id']; ?>">Delete</a>
@@ -67,14 +67,16 @@
 					</div>
 				</div> 
 			<?php } ?>
-
+<?php } ?>
 	</div> 
 
  </article>
 
 
 <footer>
+	<?php if(isset($_SESSION['username'])) { ?>
 	<a href="new.php" class="new" title="Add new student's information">Add New Entry</a>
+	<?php } ?>
 	<a href="secondary.php" class="done-button" title="Done with modify. Back to original page.">Done</a>
 </footer>
